#13 serial functions baseclassed
