+++ test/config/devices.json
@@ -2,7 +2,6 @@
     "furnace": {
         "device_type": "KDC101",
         "port": "/dev/ttyUSB0",
-        "bay_system": false,
         "stages": {
             "stage_type": "MTS50-Z8",
             "upper_limit": 10,
